Migrate fjs/bnf/data to .f.mjs, split type-level API into types.ts
proof.f.ts stays TypeScript for now: it has a runtime dependency on
testlib.f.ts and media/json/module.f.ts, both still unmigrated. Updates
descent/ll1 dependents and the stale line references across bnf/data's
todo docs.
