+++ fjs/bnf/descent/module.f.ts
@@ -18,7 +18,8 @@ import { rangeDecode } from '../module.f.mjs'
 import type { TerminalRange } from '../types.ts'
 import { contains as rangeContains } from '../../types/range/module.f.mjs'
 import { definedEntries } from '../../types/object/module.f.mjs'
-import { emptyTagMap, toData, type Rule as DataRule, type Sequence } from '../data/module.f.ts'
+import { emptyTagMap, toData } from '../data/module.f.mjs'
+import type { Rule as DataRule, Sequence } from '../data/types.ts'
 import type { Rule as FRule } from '../types.ts'
 
 export type AstTag = string|true|undefined
