+++ fjs/bnf/ll1/proof.f.ts
@@ -2,7 +2,8 @@ import { stringToCodePointList } from '../../text/utf16/module.f.mjs'
 import { toArray } from '../../types/list/module.f.mjs'
 import { commaJoin0Plus, option, range, repeat0Plus, set } from '../module.f.mjs'
 import { deterministic } from '../testlib.f.ts'
-import { type RuleSet, toData } from '../data/module.f.ts'
+import { toData } from '../data/module.f.mjs'
+import type { RuleSet } from '../data/types.ts'
 import { dispatchMap, type MatchResult, parser, parserRuleSet } from './module.f.ts'
 import { assertEq } from '../../asserts/module.f.mjs'
