+++ fjs/bnf/todo/207.md
@@ -59,7 +59,7 @@ The combinators (`option`, `repeat0Plus`, `repeat1Plus`, `join0Plus`, …) are
 just helpers that build `Variant`/`Sequence` trees, e.g. `option(x) = { some:
 x, none: [] }` and `repeat0Plus(x) = () => option([x, self])`.
 
-`parser`/`descentParser` (`fjs/bnf/data/module.f.ts`) produce a generic AST:
+`parser`/`descentParser` (`fjs/bnf/data/module.f.mjs`) produce a generic AST:
 
 ```ts
 type AstRule = { readonly tag: AstTag, readonly sequence: AstSequence }
