+++ fjs/bnf/todo/46.md
@@ -7,7 +7,7 @@ Implement an LR(1) parser because LL(1) can't handle break lines in comments.
 
 The parser targets the current BNF: functional form in
 [../module.f.mjs](../module.f.mjs), serializable form in
-[../data/module.f.ts](../data/module.f.ts). It needs an AST structure derived
+[../data/module.f.mjs](../data/module.f.mjs). It needs an AST structure derived
 from that BNF definition — see
 [parser-structure](./parser-structure.md) for the AST shape.
